Rename: slack_client.py to client.py
